pyproject.toml
[project]
name = "nyaya-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types>=0.7.0",
    "attrs>=25.3.0",
    "cachetools>=4.0",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "dataclasses-json>=0.6.7",
    "faiss-cpu>=1.11.0",
    "filelock>=3.18.0",
    "fsspec>=2025.5.1",
    "groq>=0.28.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface-hub>=0.33.0",
    "idna>=3.10",
    "joblib>=1.5.1",
    "jsonschema>=4.24.0",
    "jsonschema-specifications>=2025.4.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-text-splitters>=0.3.8",
    "langsmith>=0.3.45",
    "marshmallow>=3.26.1",
    "mypy-extensions>=1.1.0",
    "networkx>=3.5",
    "numpy>=2.3.0",
    "packaging>=24.2",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "protobuf==3.20",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.9.1",
    "pymupdf4llm>=0.0.25",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "referencing>=0.36.2",
    "regex>=2024.11.6",
    "requests>=2.32.4",
    "rpds-py>=0.25.1",
    "safetensors>=0.5.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "sentence-transformers>=4.1.0",
    "six>=1.17.0",
    "sqlalchemy>=2.0.41",
    "streamlit==1.38",
    "tenacity<9",
    "threadpoolctl>=3.6.0",
    "tiktoken>=0.9.0",
    "tokenizers>=0.21.1",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "typing-extensions>=4.14.0",
    "typing-inspect>=0.9.0",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "bs4>=0.0.2",
    "langchain-ollama>=0.3.3",
]
